Reorganized the code structure
Added Header Files and Source Files Folders. Moved all header files and source files into Header Files and Source Files folders respectively.

Changed the include paths from file name to relative paths accordingly in all the necessary files.

Note: All the commits made before this describing header and source files being stored in Header Files and Source Files folders were actually being stored in Header Files and Source Files filters.
