+++ Pokemon/Header Files/Battle/BattleManager.h
