+++ Pokemon/Header Files/Battle/BattleState.h
