+++ Pokemon/Header Files/Pokemon/WildPokemonEncounterManager.h
