+++ Pokemon/Header Files/Utility/UtilityFunctions.h
