+++ Pokemon/Pokemon.vcxproj
@@ -127,27 +127,27 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="BattleManager.cpp" />
-    <ClCompile Include="Game.cpp" />
-    <ClCompile Include="main.cpp" />
-    <ClCompile Include="Player.cpp" />
-    <ClCompile Include="Pokemon.cpp" />
-    <ClCompile Include="ProfessorOak.cpp" />
-    <ClCompile Include="UtilityFunctions.cpp" />
-    <ClCompile Include="WildPokemonEncounterManager.cpp" />
+    <ClCompile Include="Source Files\Battle\BattleManager.cpp" />
+    <ClCompile Include="Source Files\Game\Game.cpp" />
+    <ClCompile Include="Source Files\main.cpp" />
+    <ClCompile Include="Source Files\Player\Player.cpp" />
+    <ClCompile Include="Source Files\Pokemon\Pokemon.cpp" />
+    <ClCompile Include="Source Files\Character\ProfessorOak.cpp" />
+    <ClCompile Include="Source Files\Utility\UtilityFunctions.cpp" />
+    <ClCompile Include="Source Files\Pokemon\WildPokemonEncounterManager.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="BattleManager.h" />
-    <ClInclude Include="BattleState.h" />
-    <ClInclude Include="Game.h" />
-    <ClInclude Include="Grass.h" />
-    <ClInclude Include="Player.h" />
-    <ClInclude Include="Pokemon.h" />
-    <ClInclude Include="PokemonOptions.h" />
-    <ClInclude Include="PokemonType.h" />
-    <ClInclude Include="ProfessorOak.h" />
-    <ClInclude Include="UtilityFunctions.h" />
-    <ClInclude Include="WildPokemonEncounterManager.h" />
+    <ClInclude Include="Header Files\Battle\BattleManager.h" />
+    <ClInclude Include="Header Files\Battle\BattleState.h" />
+    <ClInclude Include="Header Files\Game\Game.h" />
+    <ClInclude Include="Header Files\Environment\Grass.h" />
+    <ClInclude Include="Header Files\Player\Player.h" />
+    <ClInclude Include="Header Files\Pokemon\Pokemon.h" />
+    <ClInclude Include="Header Files\Pokemon\PokemonOptions.h" />
+    <ClInclude Include="Header Files\Pokemon\PokemonType.h" />
+    <ClInclude Include="Header Files\Character\ProfessorOak.h" />
+    <ClInclude Include="Header Files\Utility\UtilityFunctions.h" />
+    <ClInclude Include="Header Files\Pokemon\WildPokemonEncounterManager.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
