+++ Pokemon/Source Files/Player/Player.cpp
@@ -1,6 +1,6 @@
-#include "Player.h"
-#include "PokemonOptions.h"
-#include "PokemonType.h"
+#include "..\..\..\..\..\GitHub\Pokemon\Pokemon\Header Files\Player\Player.h"
+#include "..\..\..\..\..\GitHub\Pokemon\Pokemon\Header Files\Pokemon\PokemonOptions.h"
+#include "..\..\..\..\..\GitHub\Pokemon\Pokemon\Header Files\Pokemon\PokemonType.h"
 
 
 Player::Player()
