+++ Pokemon/Source Files/Pokemon/WildPokemonEncounterManager.cpp
@@ -1,8 +1,8 @@
-#include "Grass.h"
-#include "Pokemon.h"
+#include "..\..\..\..\..\GitHub\Pokemon\Pokemon\Header Files\Environment\Grass.h"
+#include "..\..\..\..\..\GitHub\Pokemon\Pokemon\Header Files\Pokemon\Pokemon.h"
+#include "..\..\..\..\..\GitHub\Pokemon\Pokemon\Header Files\Pokemon\WildPokemonEncounterManager.h"
 #include <cstdlib>
 #include <ctime>
-#include "WildPokemonEncounterManager.h"
 
 WildPokemonEncounterManager::WildPokemonEncounterManager()
 {
